refactor(blog): polish AI review/brainstorm panels
- Extract StreamingChatPanel abstract base class (DRY ~80% shared code)
- Move AI Gateway URL to AI_GATEWAY_URL env var (remove hardcoded account ID)
- Remove nodejs_compat flag (Anthropic SDK removed)
- Replace regex markdown renderer with markdown-it
- Add retry button on stream failures
- Add copy button on assistant messages (hover reveal)
- Truncate API history to 20 messages (keep last 16) to avoid context window limits
- Add 30-day TTL cleanup for stale conversations in migrate script
