+++ apps/blog/wrangler.toml
@@ -4,7 +4,6 @@
 name = "maneki-blog"
 compatibility_date = "2024-12-01"
 pages_build_output_dir = "dist"
-compatibility_flags = ["nodejs_compat"]
 
 [[r2_buckets]]
 binding = "IMAGES_BUCKET"
