+++ packages/foundation/assets/subset-icons.py
@@ -46,6 +46,8 @@ CODEPOINTS: dict[str, int] = {
     "chevron_left": 0xE5CB,
     "chevron_right": 0xE5CC,
     "close": 0xE5CD,
+    "close_fullscreen": 0xF1CF,
+    "content_copy": 0xE14D,
     "download": 0xF090,
     "delete": 0xE872,
     "delete_sweep": 0xE16C,
@@ -62,6 +64,7 @@ CODEPOINTS: dict[str, int] = {
     "more_horiz": 0xE5D3,
     "more_vert": 0xE5D4,
     "notifications": 0xE7F4,
+    "open_in_full": 0xF1CE,
     "person": 0xE7FD,
     "photo_album": 0xE411,
     "progress_activity": 0xE9D0,
